pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "google-genai==1.12.1",
    "ipdb>=0.13.13",
    "python-dotenv==1.1.0",
]
